+++ Sources/Utilities/Message+Style.swift
@@ -14,11 +14,6 @@ public enum ALKMessageStyle {
         text: .text(.gray9B)
     )
 
-    public static var message = Style(
-        font: UIFont.font(.normal(size: 14)),
-        text: .text(.black00)
-    )
-
     public static var playTime = Style(
         font: UIFont.font(.normal(size: 16)),
         text: .text(.black00)
@@ -28,6 +23,29 @@ public enum ALKMessageStyle {
         font: UIFont.font(.italic(size: 12)),
         text: .text(.grayCC)
     )
+
+    // Received message text style
+    public static var receivedMessage = Style(
+        font: UIFont.font(.normal(size: 14)),
+        text: .text(.black00)
+    )
+
+    // Sent message text style
+    public static var sentMessage = Style(
+        font: UIFont.font(.normal(size: 14)),
+        text: .text(.black00)
+    )
+
+    @available(*, deprecated, message: "Use `receivedMessage` and `sentMessage`")
+    public static var message = Style(
+        font: UIFont.font(.normal(size: 14)),
+        text: .text(.black00)
+    ) {
+        didSet {
+            receivedMessage = message
+            sentMessage = message
+        }
+    }
 
     public enum BubbleStyle {
         case edge